# swagger.yaml
openapi: 3.1.0
info:
  title: Resi's Project Web Course API
  description: This is Resis's Project Web Course API docs
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: https://staging.akuanalis.com/api/v1
  - url: https://dev-api.akuanalis.com/api/v1
  - url: https://akuanalis.com/api/v1
...
paths:
...
  /users/auth/data:
    post:
      tags:
        - auth
      summary: To complete user data
      operationId: completeUserData
      security:
        - userSessionCookie: []
      requestBody:
        description: User data form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDataForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: invalid data
    put:
      tags:
        - auth
      summary: To update user data
      operationId: updateUserData
      security:
        - userSessionCookie: []
      requestBody:
        description: User data form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDataForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: invalid data
...
  /courses:
    post:
      tags:
        - course
      summary: Create a new course by admin(regular & super)
      operationId: createCourse
      security:
        - adminSessionCookie: []
      requestBody:
        description: Course object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Invalid input
    get:
      tags:
        - course
      summary: Get all courses
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
      operationId: getCourses
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponsePagination'
...
  /courses/{courseId}:
    get:
      tags:
        - course
      summary: Get course by ID
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          description: ID of the course to get
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Invalid course ID
    put:
      tags:
        - course
      summary: Update course by ID
      operationId: updateCourseById
      security:
        - adminSessionCookie: []
      parameters:
        - name: courseId
          in: path
          description: ID of the course to update
          required: true
          schema:
            type: string
      requestBody:
        description: Course object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Invalid input
    delete:
      tags:
        - course
      summary: Delete course by ID
      operationId: deleteCourseById
      security:
        - adminSessionCookie: []
      parameters:
        - name: courseId
          in: path
          description: ID of the course to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid course ID
...
  /user/{userId}/courses:
    get:
      tags:
        - user course
      summary: Get all student courses by user
      operationId: getUserCourses
      parameters:
        - name: status
          in: query
          description: status of the course
          required: true
          schema:
            type: string
            enum:
              - active
              - expired
              - all
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
      security:
        - userSessionCookie: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponsePagination'
...
components:
...
  schemas:
...
    User:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        birth:
          type: string
          format: date-time
        study_level:
          type: string
          examples:
            - sma
          enum:
            - sd
            - smp
            - sma
            - universitas
            - umum
        institution:
          type: string
          examples:
            - SMAN 1 Bekasi
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      xml:
        name: admin
    Instructor:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        education:
          type: string
          examples:
            - Bachelor of Computer Science
        experience:
          type: string
          examples:
            - 5 years
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
      xml:
        name: instructor
...
    CourseCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          examples:
            - Web Development
        description:
          type: string
          examples:
            - Learn web development from basics to advanced
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
      xml:
        name: course_category
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        admin_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          examples:
            - Web Development
        description:
          type: string
          examples:
            - Learn how to build a website
        course_category_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        purpose:
          type: string
          examples:
            - Learn how to build a website
        price:
          type: number
          examples:
            - 100000
        thumbnail_image:
          type: string
          examples:
            - adslfkad.jpg
        rating:
          type: number
          examples:
            - 4.5
        instructor_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
      xml:
        name: course
...
    CourseForm:
      type: object
      properties:
        name:
          type: string
          examples:
            - Web Development
        description:
          type: string
          examples:
            - Learn how to build a website
        category_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        price:
          type: number
          examples:
            - 100000
        purpose:
          type: string
          examples:
            - Learn how to build a website
        instructor_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
...
    UserDataForm:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        birth:
          type: string
          format: date-time
          examples:
            - 1990-01-01T00:00:00Z
        study_level:
          type: string
          examples:
            - sma
          enum:
            - sd
            - smp
            - sma
            - universitas
            - umum
        institution:
          type: string
          examples:
            - SMAN 1 Bekasi
...
    CourseResponse:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/Course'
        course_category:
          $ref: '#/components/schemas/CourseCategory'
        instructor:
          $ref: '#/components/schemas/InstructorResponse'
    CourseResponsePagination:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseResponse'
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
    InstructorResponse:
      type: object
      properties:
        instructor:
          $ref: '#/components/schemas/Instructor'
        full_name:
          type: string
          examples:
            - Jhon Doe
        photo_profile:
          type: string
          examples:
            - adslfkad.jpg
...
    PaginationResponse:
      type: object
      properties:
        page:
          type: number
          examples:
            - 1
        total_page:
          type: number
          examples:
            - 10
...
    ApiResponse:
      type: object
      properties:
        message:
          type: string
      xml:
        name: '##default'
...
  securitySchemes:
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: session
    userSessionCookie:
      type: apiKey
      in: cookie
      name: session
    instructorSessionCookie:
      type: apiKey
      in: cookie
      name: session